Refactored parser to not care about anything other than grammar.
This mainly consisted of stripping out the scope stack.  There is still
a scope stack in the parser, but it is strictly for tracking the existence
of const functions, which is necessary for parsing unary and binary
function calls correctly.

WARNING: it's not at all clear to me that the compiler's behavior is correct
anymore.  But it does compile.

Next step is to reintroduce the bits that were stripped out of the parser,
but now as a separate pass after the parser.